chore(docz-core): move test helpers out of test folder
